added router

+++ packages/core/identity/src/conn_id.rs
@@ -1,1 +1,1 @@
-pub type ConnId = u32;
+pub type ConnId = u64;
